[fix] directory structur , seperated header files
srcs/terminal
srcs/redirection
srcs/main.c

norm V3 done.
