+++ srcs/utils/make_struct.c
@@ -6,7 +6,7 @@
 /*   By: jaekpark <jaekpark@student.42seoul.fr      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/02 19:37:36 by jaekpark          #+#    #+#             */
-/*   Updated: 2021/07/25 01:47:39 by parkjaekw        ###   ########.fr       */
+/*   Updated: 2021/07/26 21:02:00 by parkjaekw        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -86,4 +86,49 @@ void			make_env(t_lst *lst, char *key, char *value)
 		tmp->next = node;
 		lst->tail = node;
 	}
+}
+
+void	make_redir(t_lst *redir, char type, char *arg)
+{
+	t_redirect *node;
+	t_redirect *tmp;
+
+	node = malloc(sizeof(t_redirect));
+	node->type = type;
+	node->arg = ft_strdup(arg);
+	node->fd = -1;
+	node->prev = NULL;
+	node->next = NULL;
+	if (redir->head == NULL && redir->tail == NULL)
+	{
+		redir->head = node;
+		redir->tail = node;
+	}
+	else if (redir->head != NULL && redir->tail != NULL)
+	{
+		tmp = redir->tail;
+		tmp->next = node;
+		node->prev = tmp;
+		redir->tail = node;
+	}
+}
+
+void	make_cmd(t_cmd *node, char *cmd, char *arg, char **args)
+{
+	if (cmd)
+		node->cmd = ft_strdup(cmd);
+	else
+		node->cmd = NULL;
+	if (arg)
+		node->arg_line = ft_strdup(arg);
+	else
+		node->arg_line = NULL;
+	if (args)
+		node->arg_word = args;
+	else
+		node->arg_word = NULL;
+	if (cmd)
+		free(cmd);
+	if (arg)
+		free(arg);
 }
