--- srcs/utils/print_error.c
@@ -1,7 +1,0 @@
-#include "../includes/pipe.h"
-
-void	print_error(char *str)
-{
-	perror(str);
-	exit(1);
-}
